update models and controllers for sql injection mitigation.

+++ app/controllers/declined_requests_controller.rb
@@ -4,14 +4,12 @@ class DeclinedRequestsController < ApplicationController
 
     def index
         @q = User.ransack(params[:q])
-        @users = @q.result
-
+        @users = @q.result.where(status: "Declined").order(created_at: :desc)
+      
         @per_page = (params[:per_page] || 10).to_i
         @page = params[:page] || 1
-
-        @requests =  @users.where(status: "Declined")
-                           .order(created_at: :desc)
-                           .paginate(page: @page, per_page: @per_page)
+      
+        @requests = @users.paginate(page: @page, per_page: @per_page)
     end
 
     def show
@@ -19,21 +17,23 @@ class DeclinedRequestsController < ApplicationController
     end
 
     def destroy
-        @user = User.find(params[:id])
-        
-        if @user.destroy
-          redirect_to declined_requests_path, notice: "You successfully deleted #{@user.email}'s application."
+        @user = User.find_by(id: params[:id])
+        if @user.nil?
+          redirect_to declined_requests_path, alert: "User not found."
+        else
+
+            if @user.destroy
+                redirect_to declined_requests_path, notice: "You successfully deleted #{@user.email}'s application."
+            else
+                redirect_to declined_requests_path, alert: "Failed to delete user. Please try again."
+            end
         end
     end
 
     private
 
     def verify_is_admin
-        if current_user.admin?
-           return
-        else
-           redirect_to root_path, notice: "You must be an admin to perform this action."
-        end
+        redirect_to root_path, notice: "You must be an admin to perform this action." unless current_user.admin?
     end
 
 end
